+++ src/lib-tempo/CMakeLists.txt
@@ -13,7 +13,7 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentID.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentPlayerLocal.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentPlayerRemote.hpp"
-	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/ComponentPlayerRemoteServer.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentPlayerRemoteServer.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentRender.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentTransform.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/EntityCreation.hpp"
@@ -31,7 +31,7 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemLevelManager.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemPlayerLocal.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemPlayerRemote.hpp"
-	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/SystemPlayerRemoteServer.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemPlayerRemoteServer.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemRender.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemRenderHealth.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemTransform.hpp"
@@ -49,7 +49,7 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentID.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentPlayerLocal.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentPlayerRemote.cpp"
-	"${CMAKE_CURRENT_LIST_DIR}/src/entity/ComponentPlayerRemoteServer.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentPlayerRemoteServer.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentRender.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentTransform.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/EntityCreation.cpp"
@@ -67,7 +67,7 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemLevelManager.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemPlayerLocal.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemPlayerRemote.cpp"
-	"${CMAKE_CURRENT_LIST_DIR}/src/entity/SystemPlayerRemoteServer.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemPlayerRemoteServer.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemRender.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemRenderHealth.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemTransform.cpp"
