+++ src/lib-tempo/src/component/ComponentPlayerRemoteServer.cpp
@@ -1,0 +1,7 @@
+#include <tempo/entity/SystemPlayerRemoteServer.hpp>
+
+namespace tempo{
+	
+	//Nothing Implemented
+
+}
